updated scripts

+++ utility/genome_bed_shuffle_v3.py
@@ -1,29 +1,44 @@
 import sys, math, glob, multiprocessing, subprocess, os, bisect, random, re
 import pysam, pybedtools
 
-# Usage: python genome_bed_shuffle.py [-h] [-q] [-t=gc_threshold] [-s=seed_val] [-n=max_tries] [-o=out_id] [-p=pos_file] [-r=pos_threshold] <fasta_file> <bed_file>
-NTRIES=1000
-THRESH=0.05
-POSTHRESH = -1
+# Usage: python genome_bed_shuffle_v3.py [-h] [-q] [-k] [-s=seed_val] 
+# [-n=max_tries] [-o=out_id] [-t=pos_file] [-p=pos_threshold] [-b=bam_files]
+# [-r=min_read_cov] [-i=include_file] <fasta_file> <bed_file>
 
-def processInputs( bedFileStr, fastaFileStr, posFileStr, maxTries, gcThresh, outId, seedVal, posThresh, isPrint ):
+NTRIES=1000
+POSTHRESH = -1
+NREADS=5
+
+def processInputs( bedFileStr, fastaFileStr, posFileStr, bamFileStr, minReads, maxTries, outId, seedVal, posThresh, includeFileStr, isKeep, isPrint ):
 
 	isPos = (posFileStr != None)
-	
+	isBam = (bamFileStr != None)
 	
 	if outId == None:
 		tmp = os.path.basename( bedFileStr )
 		outId = re.sub( "\.bed", "", tmp, flags=re.IGNORECASE )
 	outFileStr = outId + '_equiv' + '.bed'
+	
+	if isBam:
+		bamFileAr = readBamList( bamFileStr )
+	
+	# check include file
+	if includeFileStr != None and os.path.isfile(includeFileStr) == False:
+		print( 'WARNING: include file {:s} does not exist...ignoring it'.format(os.path.basename(includeFileStr)))
+		includeFileStr = None
 		
 	if isPrint:
 		print( 'Max Tries:', maxTries )
-		print( 'GC Threshold:', gcThresh )
 		print( 'Fasta file:', os.path.basename(fastaFileStr) )
 		if isPos:
 			print( 'Position Threshold:', posThresh)
 			print( 'Position file:', os.path.basename(posFileStr) )
 		print( 'Starting seed:', seedVal )
+		if isBam:
+			print( 'BAM files:', len(bamFileAr) )
+			print( 'Min read coverage:', minReads )
+		if includeFileStr != None:
+			print( 'Include regions file:', os.path.basename(includeFileStr) )
 		print( 'BED file:', os.path.basename(bedFileStr) )
 		print( 'Out file:', os.path.basename(outFileStr))
 	
@@ -41,19 +56,21 @@ def processInputs( bedFileStr, fastaFileStr, posFileStr, maxTries, gcThresh, out
 	
 	# build position dict if necessary
 	if isPos:
+		if isPrint:
+			print('Getting positions')
 		posDict = readPosFile( posFileStr )
 	else:
 		posDict = None
 	
+	#updatedDmrs = dmrFile.each( computeCGDens, fasta=fastaFile )
 	if isPrint:
-		print( 'Getting regions and nucleotide density' )
-	
-	updatedDmrs = dmrFile.each( computeCGDens, fasta=fastaFile )
+		print( dmrFile.count(), 'features in input file' )
+	updatedDmrs = dmrFile
 	
 	if isPrint:
 		print('Running process')
 		
-	nTries, lastFile, leftOverDMRs = runProcess( maxTries, updatedDmrs, fastaFile, fastaIndex, gcThresh, posDict, posThresh, seedVal, isPrint )
+	nTries, lastFile, leftOverDMRs = runProcess( maxTries, updatedDmrs, fastaFile, fastaIndex, posDict, posThresh, bamFileAr, minReads, seedVal, includeFileStr, isPrint )
 	if nTries >= maxTries:
 		print('WARNING: Did not converge')
 		leftOverDMRs.saveas( outId + '_unconverged.bed')
@@ -62,12 +79,32 @@ def processInputs( bedFileStr, fastaFileStr, posFileStr, maxTries, gcThresh, out
 	
 	pybedtools.BedTool(lastFile).sort(faidx=fastaIndex).moveto(outFileStr)
 	
-	tmpFileAr = glob.glob('tmp*')
-	#print(tmpFileAr)
-	for x in tmpFileAr:
-		os.remove(x)
+	if not isKeep:
+		tmpFileAr = glob.glob('tmp*')
+		for x in tmpFileAr:
+			os.remove(x)
 	if isPrint:
 		print( 'Done.' )
+
+def readBamList( bamListFile ):
+	inFile = open( bamListFile, 'r' )
+	outAr = []
+	
+	for line in inFile:
+		if line.startswith('#'):
+			continue
+		bamFileStr = line.rstrip()
+		bamIndex = bamFileStr + '.bai'
+		if os.path.isfile(bamFileStr) == False:
+			print( 'WARNING: BAM file {:s} does not exist...skipping'.format( os.path.basename(bamFileStr) ) )
+			continue
+		if os.path.isfile(bamIndex) == False:
+			print( 'Creating BAM index for', os.path.basename(bamFileStr))
+			pysam.index( bamFileStr )
+		outAr += [bamFileStr]
+	# end for line
+	inFile.close()
+	return outAr
 	
 def checkFasta( fastaFileStr ):
 	fastaIndex = fastaFileStr + '.fai'
@@ -112,28 +149,13 @@ def readPosFile( posFileStr ):
 	# end for line
 	posFile.close()
 	return outDict
-	
-def computeCGDens( feature, fasta=None ):
-	if fasta == None:
-		score = -1
-	else:
-		score = _computeCG( fasta, feature.chrom, feature.start, feature.stop )
-	feature.score = str(score)
-	return feature
 
-def _computeCG( fastaFile, chrm, start, end ):
-	refSeq = fastaFile.fetch( chrm, start, end)
-	
-	cgList = list(filter(lambda r: r == 'C' or r == 'G', refSeq))
-	return len(cgList) / float(end - start)
-
-def runProcess( maxTries, updatedDmrs, fastaFile, fastaIndex, gcThresh, posDict, posThresh, seedVal, isPrint ):
+def runProcess( maxTries, updatedDmrs, fastaFile, fastaIndex, posDict, posThresh, bamFileAr, minReads, seedVal, includeFileStr, isPrint ):
 	
 	excludeFileStr = 'tmp.exclude.bed'
 	outFileStr = 'tmp.equiv_-1.bed'
 	remainFileStr = 'tmp.original.bed'
 	remainingDMRs = updatedDmrs.saveas(remainFileStr)
-	#outBedFile = 'equiv_regions.bed'
 	
 	foundRegions = pybedtools.BedTool('', from_string=True).saveas(outFileStr)
 	
@@ -142,7 +164,7 @@ def runProcess( maxTries, updatedDmrs, fastaFile, fastaIndex, gcThresh, posDict,
 		x = foundRegions.cat('tmp.original.bed', force_truncate=True).moveto(excludeFileStr)
 		
 		# try to find intervals
-		tryAgainDMRs, newRegions = runTry(i, gcThresh, seedVal, remainFileStr, excludeFileStr, fastaFile, fastaIndex, posDict, posThresh, isPrint )
+		tryAgainDMRs, newRegions = runTry(i, seedVal, remainFileStr, excludeFileStr, fastaFile, fastaIndex, posDict, posThresh, bamFileAr, minReads, includeFileStr, isPrint )
 		
 		#  add new found regions to old
 		
@@ -161,52 +183,63 @@ def runProcess( maxTries, updatedDmrs, fastaFile, fastaIndex, gcThresh, posDict,
 	# end for i
 	return i, outFileStr, remainingDMRs
 	
-def runTry( i, gcThresh, seedVal, remainFileStr, excludeFileStr, fastaFile, fastaIndex, posDict, posThresh, isPrint ):
+def runTry( i, seedVal, remainFileStr, excludeFileStr, fastaFile, fastaIndex, posDict, posThresh, bamFileAr, minReads, includeFileStr, isPrint ):
 	
-	isGC = gcThresh != -1
 	isPos = posThresh != -1
+	isBam = len(bamFileAr) > 0
+	
+	if isBam:
+		bamPointerAr = [ pysam.AlignmentFile( bamFileStr, 'rb' ) for bamFileStr in bamFileAr ]
+		
 	remainDMRs = pybedtools.BedTool(remainFileStr)
-	#if isPrint:
-		#print( 'Iteration {:d} with {:d} regions...'.format(i, remainDMRs.count()), end='' )
+
 	# generate shuffle of remaining DMRs
-	if seedVal == None:
-		#outBed = remainDMRs.shuffle( g=fastaIndex, excl=excludeFileStr, chrom=True )
+	if seedVal == None and includeFileStr == None:
 		outBed = remainDMRs.shuffle( g=fastaIndex, excl=excludeFileStr)
-	else:
-		#outBed = remainDMRs.shuffle( g=fastaIndex, excl=excludeFileStr, chrom=True, seed=seedVal )
+	elif seedVal == None:
+		outBed = remainDMRs.shuffle( g=fastaIndex, excl=excludeFileStr, incl=includeFileStr )
+	elif includeFileStr == None:
 		seedVal += i
 		outBed = remainDMRs.shuffle( g=fastaIndex, excl=excludeFileStr, seed=seedVal )
+	else:
+		seedVal += i
+		outBed = remainDMRs.shuffle( g=fastaIndex, excl=excludeFileStr, incl=includeFileStr, seed=seedVal )
 	
 	# iterate to see which to keep
-	#acceptDMR = [] # DMR labels
 	tryAgainDMR = [] # DMR labels
 	outAr = [] # output regions
 	
 	for feature in outBed:
+		nRead = 0
 		isValid = True
 		# if close enough
 		#print('  ', cgDens, feature.score, abs(cgDens - float(feature.score) ) <= THRESH)
-		if not isGC and not isPos: # no thresholds
+		if not isPos and not isBam: # no thresholds
 			outAr += [ feature ]
 			continue
-		
-		if isGC: # check gc content if necessary
-			cgDens = _computeCG( fastaFile, feature.chrom, feature.start, feature.stop )
-			isValid = (abs(cgDens - float(feature.score) ) <= gcThresh)
-		
+
 		if isPos: # also check position
 			posAr = posDict.get( feature.chrom )
 			nPos = _computePos( posAr, feature.start, feature.stop )
 			#print(feature.name, nPos)
 			isValid = isValid and (nPos >= posThresh)
 		
+		if isBam: # check read coverage
+			nReadAr = [ bamFile.count(feature.chrom, feature.start, feature.stop) for bamFile in bamPointerAr ]
+			nRead = min ( nReadAr )
+			isValid = isValid and (nRead >= minReads)
+		
 		if isValid:
 			#acceptDMR += [feature.label]
+			feature.score = nRead
 			outAr += [ feature ]
 		else:
 			 tryAgainDMR += [ feature.name ]
 	# end for feature
-
+	# close bam files
+	if isBam:
+		t = [ x.close() for x in bamPointerAr ]
+		
 	if isPrint:
 		print('{:d} regions remaining'.format( len(tryAgainDMR)) )
 	return tryAgainDMR, outAr
@@ -223,7 +256,13 @@ def _computePos( posAr, start, end ):
 		return 0
 	# how many positions are there
 	return eIndex - sIndex
+
+'''def readCoverage( bamPointerAr, chrm, start, end ):
+	outAr = []
 	
+	for bamFile in bamPointerAr:
+		nReads = bamFile.count( chrm, start, end )'''
+		
 def bisect_ge(a, x):
 	# leftmost (first) INDEX greater than or equal to x
 	# return None if not found
@@ -242,17 +281,23 @@ def bisect_lt(a, x):
 	
 def parseInputs( argv ):
 	maxTries = NTRIES
-	gcThresh = THRESH
 	posThresh = POSTHRESH
+	minReads = NREADS
+	bamFileList = None
 	outId = None
 	seedVal = None
 	posFileStr = None
+	includeFileStr = None
 	isPrint = True
+	isKeep = False
 	startInd = 0
 	
-	for i in range(min(7, len(argv))):
+	for i in range(min(9, len(argv))):
 		if argv[i] == '-q':
 			isPrint = False
+			startInd += 1
+		elif argv[i] == '-k':
+			isKeep = True
 			startInd += 1
 		elif argv[i].startswith('-n='):
 			try:
@@ -266,25 +311,29 @@ def parseInputs( argv ):
 			except ValueError:
 				print( 'WARNING: seed value must be an integer...using None')
 			startInd += 1
-		elif argv[i].startswith('-p='):
+		elif argv[i].startswith('-t='):
 			posFileStr = argv[i][3:]
 			startInd += 1
-		elif argv[i].startswith('-r='):
+		elif argv[i].startswith('-p='):
 			try:
 				posThresh = int(argv[i][3:])
 			except ValueError:
 				print( 'WARNING: position threshold must be an integer...using None' )
 			startInd += 1
-		elif argv[i].startswith('-t='):
+		elif argv[i].startswith('-r='):
 			try:
-				gcThresh = float(argv[i][3:])
-				if gcThresh >= 1:
-					gcThresh = gcThresh / 100.0
+				minReads = int(argv[i][3:])
 			except ValueError:
-				print( 'WARNING: GC threshould must be a float...using default', THRESH)
+				print( 'WARNING: min read coverage must be an interger...using default', NREADS)
 			startInd += 1
 		elif argv[i].startswith('-o='):
 			outId = argv[i][3:]
+			startInd += 1
+		elif argv[i].startswith('-b='):
+			bamFileList = argv[i][3:]
+			startInd += 1
+		elif argv[i].startswith('-i='):
+			includeFileStr = argv[i][3:]
 			startInd += 1
 		elif argv[i] == '-h':
 			printHelp()
@@ -302,10 +351,12 @@ def parseInputs( argv ):
 	fastaFileStr = argv[startInd]
 	bedFileStr = argv[startInd + 1]
 	
-	processInputs( bedFileStr, fastaFileStr, posFileStr, maxTries, gcThresh, outId, seedVal, posThresh, isPrint )
+	processInputs( bedFileStr, fastaFileStr, posFileStr, bamFileList, minReads, maxTries, outId, seedVal, posThresh, includeFileStr, isKeep, isPrint )
 
 def printHelp():
-	print( 'Usage: python genome_bed_shuffle.py [-h] [-q] [-t=gc_threshold] [-s=seed_val] [-n=max_tries] [-o=out_id] [-p=pos_file] [-r=pos_threshold] <fasta_file> <bed_file>' )
+	print( 'python\tgenome_bed_shuffle_v3.py [-h] [-q] [-k] [-s=seed_val]' )
+	print( '\t[-n=max_tries] [-o=out_id] [-t=pos_file] [-p=pos_threshold] [-b=bam_files]')
+	print('\t[-r=min_read_cov] [-i=include_file]<fasta_file> <bed_file>')
 	print('WARNING: temporary files are written using predefined names. DO NOT run multiple instances of this script in the same directory at the same time.')
 
 if __name__ == "__main__":
